chore: consolidate line cursor code
